moving home screen components to specific package
